# 🎯 Complete Architecture Consolidation + Mobile Menu Enhancements + Critical Bug Fixes