--- a/client/src/components/GameHeader.tsx
+++ b/client/src/components/GameHeader.tsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import { useGameStore } from '../stores';
+import { useGameStore, useUIStore } from '../stores';
 
 interface GameHeaderProps {
   onBackToMenu?: (() => void) | undefined;
@@ -7,10 +7,16 @@ interface GameHeaderProps {
 
 export function GameHeader({ onBackToMenu }: GameHeaderProps) {
   const [isMobile, setIsMobile] = useState(false);
-  const [isLeaving, setIsLeaving] = useState(false);
-  const roomCode = useGameStore(state => state.roomCode);
-  const connectionStatus = useGameStore(state => state.connectionStatus);
-  const leaveRoom = useGameStore(state => state.leaveRoom);
+
+  const { roomCode, connectionStatus } = useGameStore(state => ({
+    roomCode: state.roomCode,
+    connectionStatus: state.connectionStatus,
+  }));
+
+  const { toggleInGameMenu, toggleRoomInfo } = useUIStore(state => ({
+    toggleInGameMenu: state.toggleInGameMenu,
+    toggleRoomInfo: state.toggleRoomInfo,
+  }));
 
   // Detect mobile screen size
   useEffect(() => {
@@ -23,24 +29,9 @@ export function GameHeader({ onBackToMenu }: GameHeaderProps) {
     return () => window.removeEventListener('resize', checkMobile);
   }, []);
 
-  const handleLeaveRoom = async () => {
-    if (isLeaving) return;
+  // Note: handleLeaveRoom is now handled by the InGameMenu component
 
-    setIsLeaving(true);
-    try {
-      const success = await leaveRoom();
-      if (success && onBackToMenu) {
-        onBackToMenu();
-      }
-    } catch (error) {
-      console.error('Error leaving room:', error);
-    } finally {
-      setIsLeaving(false);
-    }
-  };
-
-  const showLeaveButton =
-    roomCode && connectionStatus === 'connected' && !isLeaving;
+  const showLeaveButton = roomCode && connectionStatus === 'connected';
 
   return (
     <header
@@ -70,36 +61,12 @@ export function GameHeader({ onBackToMenu }: GameHeaderProps) {
           flexShrink: 0,
         }}
       >
-        {showLeaveButton ? (
-          // When in a room, show Leave Room button instead of Back button
+        {showLeaveButton || onBackToMenu ? (
+          // Show hamburger menu for in-game options
           <button
-            className={`leave-room-btn ${isMobile ? 'mobile-leave' : ''}`}
-            onClick={handleLeaveRoom}
-            disabled={isLeaving}
-            title="Leave Room"
-            style={{
-              backgroundColor: isLeaving ? '#666' : '#d32f2f',
-              color: 'white',
-              border: 'none',
-              padding: '8px 12px',
-              borderRadius: '6px',
-              cursor: isLeaving ? 'not-allowed' : 'pointer',
-              fontSize: isMobile ? '14px' : '16px',
-              fontWeight: 'bold',
-              minWidth: '40px',
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'center',
-            }}
-          >
-            {isLeaving ? '...' : '✕'}
-          </button>
-        ) : onBackToMenu ? (
-          // When not in a room, show Back to Menu button
-          <button
-            className={`back-to-menu-btn ${isMobile ? 'mobile-back' : ''}`}
-            onClick={onBackToMenu}
-            title="Back to Menu"
+            className={`hamburger-menu-btn ${isMobile ? 'mobile-hamburger' : ''}`}
+            onClick={toggleInGameMenu}
+            title="Game Menu"
             style={{
               backgroundColor: 'rgba(255, 255, 255, 0.1)',
               color: 'white',
@@ -108,13 +75,15 @@ export function GameHeader({ onBackToMenu }: GameHeaderProps) {
               borderRadius: '6px',
               cursor: 'pointer',
               fontSize: isMobile ? '14px' : '16px',
+              fontWeight: 'bold',
               minWidth: '40px',
               display: 'flex',
               alignItems: 'center',
               justifyContent: 'center',
+              transition: 'all 0.2s ease',
             }}
           >
-            ←
+            ☰
           </button>
         ) : null}
       </div>
@@ -142,16 +111,35 @@ export function GameHeader({ onBackToMenu }: GameHeaderProps) {
         </h1>
       </div>
 
-      {/* Room Info Section */}
+      {/* Room Info Section - Clickable */}
       {roomCode && (
         <div
-          className="room-info"
+          className="room-info clickable"
+          onClick={toggleRoomInfo}
+          title="Click for room details"
           style={{
             fontSize: isMobile ? '10px' : '11px',
-            opacity: 0.7,
+            opacity: 0.8,
             textAlign: 'right',
             flexShrink: 0,
             lineHeight: '1.2',
+            cursor: 'pointer',
+            padding: '4px 8px',
+            borderRadius: '4px',
+            transition: 'all 0.2s ease',
+            border: '1px solid transparent',
+          }}
+          onMouseEnter={e => {
+            const target = e.target as HTMLElement;
+            target.style.opacity = '1';
+            target.style.backgroundColor = 'rgba(255, 255, 255, 0.1)';
+            target.style.borderColor = 'rgba(255, 255, 255, 0.2)';
+          }}
+          onMouseLeave={e => {
+            const target = e.target as HTMLElement;
+            target.style.opacity = '0.8';
+            target.style.backgroundColor = 'transparent';
+            target.style.borderColor = 'transparent';
           }}
         >
           <div>{roomCode}</div>