--- a/client/src/hooks/index.ts
+++ b/client/src/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useResponsive, useIsMobile } from './useResponsive';