… FEATURES.md
… TODO.md
… client/src/components/GameInfo.tsx
… client/src/components/MenuScreen.module.css
… client/src/components/MobileOpponentArea.tsx
… client/src/components/MultiOpponentArea.tsx
… client/src/components/OpponentArea.tsx
… client/src/components/PlayerHandArea.tsx
… client/src/components/Toast.module.css
… client/src/components/Toast.tsx
… client/src/components/mobile/MobileGameBoard.tsx
… client/src/components/mobile/MobileHandArea.tsx
… client/src/components/mobile/MobileOpponentArea.tsx
… client/src/components/mobile/MobilePlayerSheet.module.css
… client/src/components/mobile/MobilePlayerSheet.tsx
… client/src/hooks/useResponsive.ts
… client/src/stores/gameStore.ts
… client/src/stores/uiStore.ts
… client/src/styles/globals.css
… client/src/utils/toastUtils.ts
… server/src/index.ts
… shared/src/game-engine/game-engine.ts
… shared/src/types/events.ts
… shared/src/types/game.ts